Description
The project consisted on moving across diferent states to provide the customer
a car wash. Those stes were basic, extra, and platimum wash.
A state machine is composed of states that perform an operations after
an event is triggered it goes to another state. In order to end up a program
there is a state called "end state" which will basically tell the system
to halt or end.

During this lab we learned how to use all the tools that we learned during
the whole semester. Some of those tools are: state machines, an external clock,
switches, buttons, and an external display to show what is the current state.

Once the lab was finished, it was amazing to see how the program automatically
changed between states according to the program that was selected. For example,
after inserting a coin, you could select the basic program. Then the display
will show a (bP) and soon after will perform the next three steps:
1. Apply Soap Foam (SF)
2. Rubbing with Mitter curtain & Scrubbers (C)
3. Rinsing (r1)
After performing the steps the machine goes back to idle and display (--).